Do not require login to view DOS lot start pages
Moved start page route from the `buyers` blueprint to the `main` blueprint, so that users do not need to log in to see them.  URLs remain the same.
