+++ app/main/__init__.py
@@ -3,4 +3,7 @@ from flask import Blueprint
 main = Blueprint('main', __name__)
 
 from . import errors
-from .views import crown_hosting, digital_services_framework, g_cloud, login, marketplace, suppliers
+from .views import (
+    crown_hosting, digital_services_framework, g_cloud, login, marketplace, suppliers,
+    digital_outcomes_and_specialists
+)
